Mostrar tipo (alternativa, comercial) Filtrar tipo Detectados y corregidos errores de codificación que podrían afectar a la integridad de la BD Consultar para modificar tipo y cluster de un articulo

+++ ProyectoWI/src/java/tienda/DAO/ArticuloDAO.java
@@ -32,12 +32,20 @@ public class ArticuloDAO {
     MySQLMetodos m = new MySQLMetodos();
 
     //ACTUALIZADA
+    /**
+     * 
+     * @param filtrarActivos
+     * @param condicion - puede ser nula, si no lo es tiene que empezar por AND.
+     * @return 
+     */
     public List<Articulo> findArticulos(boolean filtrarActivos, String condicion) {
         List<Articulo> l = new ArrayList<Articulo>();
         try {
             conexion = m.obtenerConexionDAWA();
             Statement sentenciaSQL = conexion.createStatement();
-            String query = "SELECT a.codigoArticulo, a.precio, a.unidades, a.activo, a.idPelicula, idCluster, spanishTitle as titulo, imdbPictureURL as imagen, year as anho "
+            String query = "SELECT a.codigoArticulo, a.precio, a.unidades, a.activo, "
+                    + " a.idPelicula, idCluster, tipo, spanishTitle as titulo, "
+                    + " imdbPictureURL as imagen, year as anho "
                     + " FROM `articulos` a,`movies` m "
                     + " WHERE a.idPelicula = m.id ";
             if(filtrarActivos){
@@ -47,15 +55,19 @@ public class ArticuloDAO {
             System.out.println("ArticuloDAO:" + query);
             ResultSet consulta = sentenciaSQL.executeQuery(query);
             while (consulta.next()) {
-                // TODO: obtener otros datos de la película si es necesario en la pagina web
                 Pelicula p = new Pelicula(consulta.getInt("idPelicula"),
                         consulta.getString("titulo"),
                         consulta.getInt("anho"),
                         consulta.getString("imagen"));
                 Articulo a = new Articulo(consulta.getString("codigoArticulo"), 
                         consulta.getFloat("precio"), consulta.getBoolean("activo"), 
-                        consulta.getInt("unidades"), consulta.getInt("idCluster"), p);
+                        consulta.getInt("unidades"), consulta.getInt("idCluster"), 
+                        consulta.getInt("tipo"), p);
+                if(consulta.wasNull()){
+                    a.setTipo(null);
+                }
                 l.add(a);
+                
             }
         } catch (Exception ex) {
             Logger.getLogger(ArticuloDAO.class.getName()).log(Level.SEVERE, null, ex);
@@ -92,52 +104,52 @@ public class ArticuloDAO {
         return findArticuloById(idArticulo, true, true, true, true);
     }
 
-    //TODO: PROBAR Y COMPLETAR
     public Articulo findArticuloById(String idArticulo, boolean conActores, 
             boolean conDirectores, boolean conGeneros, boolean conPaises ) {
             String condicion = " AND codigoArticulo ='" + idArticulo + "'";
-            Articulo a = findArticulos(false, condicion).get(0);
-            int movieID = a.getPelicula().getId();
-            if(conActores || conDirectores || conGeneros || conPaises){
-                PeliculasDAO peliDAO = new PeliculasDAO();
-                if(conDirectores){
-                    a.getPelicula().setDirectores(peliDAO.getDirectores(movieID));
+            try{
+                Articulo a = findArticulos(false, condicion).get(0);
+                int movieID = a.getPelicula().getId();
+                if(conActores || conDirectores || conGeneros || conPaises){
+                    PeliculasDAO peliDAO = new PeliculasDAO();
+                    if(conDirectores){
+                        a.getPelicula().setDirectores(peliDAO.getDirectores(movieID));
+                    }
+                    if(conActores){
+                        a.getPelicula().setActores(peliDAO.getActores(movieID));
+                    }
+                    if(conGeneros){
+                        a.getPelicula().setGeneros(peliDAO.getGeneros(movieID));
+                    }
+                    if(conPaises){
+                        a.getPelicula().setPaises(peliDAO.getPaises(movieID));
+                    }
                 }
-                if(conActores){
-                    a.getPelicula().setActores(peliDAO.getActores(movieID));
-                }
-                if(conGeneros){
-                    a.getPelicula().setGeneros(peliDAO.getGeneros(movieID));
-                }
-                if(conPaises){
-                    a.getPelicula().setPaises(peliDAO.getPaises(movieID));
-                }
+                return a;
+            }catch(IndexOutOfBoundsException iob){
+                
             }
-            return a;
+            return null;
     }
 
-    public List<Articulo> findArticulosFiltrados(String anho, Float precioMaximo, String titulo){
-    	boolean filtrarAnho = anho!=null && !anho.isEmpty();
-    	boolean filtrarPM = precioMaximo!=null;
-    	boolean filtrarTitulo = titulo!=null && !titulo.isEmpty();
-    	return findArticulosFiltrados(anho, filtrarAnho, precioMaximo, filtrarPM, titulo, filtrarTitulo);
-    }
-    
-    
-    public List<Articulo> findArticulosFiltrados(String anho, Boolean filtrarAnho,
-            Float precioMaximo, Boolean filtrarPM, String titulo, Boolean filtrarTitulo) {
+    public List<Articulo> findArticulosFiltrados(String anho, Float precioMaximo, 
+            String titulo, Integer tipo) {
         String condicion = "";
 
-        if (filtrarAnho && !anho.isEmpty()) {
+        if (anho != null) {
             condicion += " AND year=" + anho;
         }
 
-        if (filtrarPM) {
+        if (precioMaximo != null ) {
             condicion += " AND precio<=" + precioMaximo;
         }
 
-        if (filtrarTitulo && !titulo.isEmpty()) {
+        if (titulo!= null && !titulo.isEmpty()) {
             condicion += " AND spanishTitle like '%" + titulo + "%'";
+        }
+        
+        if (tipo != null){
+            condicion += " AND tipo =" + tipo.intValue();
         }
 
         return findArticulos(true, condicion);
@@ -147,6 +159,8 @@ public class ArticuloDAO {
     
     // actualizada
     public Boolean existeArticulo(Articulo a) {
+        if(a.getCodigoArticulo() == null || a.getCodigoArticulo().isEmpty())
+            return false;
         try {
             conexion = m.obtenerConexionDAWA();
             Statement sentenciaSQL = conexion.createStatement();
@@ -174,7 +188,6 @@ public class ArticuloDAO {
             //Falta ver que el código no está duplicado
             conexion = m.obtenerConexionDAWA();
             Statement sentenciaSQL = conexion.createStatement(ResultSet.TYPE_SCROLL_SENSITIVE, ResultSet.CONCUR_UPDATABLE);
-            //TODO corregir este insert para introducir articulos y peliculas
             String query = "INSERT INTO `articulos` (`codigoArticulo`, `precio`, `activo`, `unidades`, `idPelicula`)" 
                     + " VALUES ('" + a.getCodigoArticulo() + "', " + a.getPrecio() 
                     + ", " + a.getActivo() + ", " + a.getUnidades() + ", " + a.getPelicula().getId() + ");";
@@ -191,6 +204,8 @@ public class ArticuloDAO {
 
     //probada actualizada
     public boolean borrarArticulo(String codigoArticulo) {
+        if(codigoArticulo == null ||codigoArticulo.isEmpty())
+            return false;
         try {
             conexion = m.obtenerConexionDAWA();
             String query = "DELETE from articulos where codigoArticulo='" + codigoArticulo + "'";
@@ -207,11 +222,16 @@ public class ArticuloDAO {
 
     // actualizada
     public boolean modificarArticulo(Articulo a) {
+        if(a.getCodigoArticulo() == null ||a.getCodigoArticulo().isEmpty())
+            return false;
         try {
             conexion = m.obtenerConexionDAWA();
             String query = "UPDATE articulos "
-                    + "set precio=" + a.getPrecio() + ",activo=" + a.getActivo() 
+                    + "set precio=" + a.getPrecio() 
+                    + ",activo=" + a.getActivo() 
                     + ",unidades=" + a.getUnidades()
+                    + ",tipo=" + a.getTipo()
+                    + ",idC"
                     + " where codigoArticulo='" + a.getCodigoArticulo() + "'";
             m.ejecutarOperacion(conexion, query);
             return true;
@@ -225,9 +245,12 @@ public class ArticuloDAO {
 
     // actualizada
     public boolean activarArticulo(String idArticulo, Boolean activar) {
+        if(activar == null || idArticulo == null || idArticulo.isEmpty()) 
+            return false;
         try {
             conexion = m.obtenerConexionDAWA();
-            String query = "UPDATE articulos set " + " activo=" + activar + "" + " where codigoArticulo='" + idArticulo + "'";
+            String query = "UPDATE articulos set " + " activo=" + activar
+                    + " where codigoArticulo='" + idArticulo + "'";
             System.out.println("query: " + query);
             m.ejecutarOperacion(conexion, query);
             return true;
@@ -239,16 +262,65 @@ public class ArticuloDAO {
         }
     }
     
-        // actualizada
-    public void modificarUnidades(Articulo a) {
+    // actualizada
+    public boolean modificarUnidades(Articulo a) {
+        if(a.getUnidades() == null || a.getCodigoArticulo() == null || a.getCodigoArticulo().isEmpty()) 
+            return false;
         try {
             conexion = m.obtenerConexionDAWA();
-            String query = "UPDATE articulos set " + " unidades=" + a.getUnidades() 
+            String query = "UPDATE articulos set unidades=" + a.getUnidades() 
                     + " where codigoArticulo='" + a.getCodigoArticulo() + "'";
             System.out.println("query: " + query);
             m.ejecutarOperacion(conexion, query);
+            return true;
         } catch (Exception ex) {
             Logger.getLogger(UsuarioDAO.class.getName()).log(Level.SEVERE, null, ex);
+            return false;
+        } finally {
+            m.cerrarConexion(conexion);
+        }
+    }
+    
+    /**
+     * 
+     * @param a - solo necesita tener fijado el codigo del articulo y el valor
+     * de cluster. Si el valor de cluster es null devuelve false y no se hace nada.
+     */
+    public boolean modificarCluster(Articulo a) {
+        if(a.getCluster() == null || a.getCodigoArticulo() == null || a.getCodigoArticulo().isEmpty()) 
+            return false;
+        try {
+            conexion = m.obtenerConexionDAWA();
+            String query = "UPDATE articulos set idCluster=" + a.getCluster() 
+                    + " where codigoArticulo='" + a.getCodigoArticulo() + "'";
+            System.out.println("query: " + query);
+            m.ejecutarOperacion(conexion, query);
+            return true;
+        } catch (Exception ex) {
+            Logger.getLogger(UsuarioDAO.class.getName()).log(Level.SEVERE, null, ex);
+            return false;
+        } finally {
+            m.cerrarConexion(conexion);
+        }
+    }
+    /**
+     * Permite modificar el tipo de un artículo 
+     * @param a - solo necesita tener fijado el id de la película y el valor
+     * de tipo (puede ser nulo).
+     */
+    public boolean modificarTipo(Articulo a) {
+        if(a.getIdPelicula() == null) 
+            throw new NullPointerException();
+        try {
+            conexion = m.obtenerConexionDAWA();
+            String query = "UPDATE articulos set tipo=" + a.getTipo() 
+                    + " where idPelicula=" + a.getIdPelicula();
+            System.out.println("query: " + query);
+            m.ejecutarOperacion(conexion, query);
+            return true;
+        } catch (Exception ex) {
+            Logger.getLogger(UsuarioDAO.class.getName()).log(Level.SEVERE, null, ex);
+            return false;
         } finally {
             m.cerrarConexion(conexion);
         }
